Add feature to set the interval between separator lines in hours

+++ library/src/main/java/com/alamkanak/weekview/DateTimeInterpreter.java
@@ -7,5 +7,5 @@ import java.util.Calendar;
  */
 public interface DateTimeInterpreter {
     String interpretDate(Calendar date);
-    String interpretTime(int hour);
+    String interpretTime(float hour);
 }
